+++ packages/@runlightyear/hubspot/src/HubSpotOAuth.ts
@@ -23,6 +23,8 @@ export class HubSpotOAuth extends OAuthConnector {
         "crm.objects.companies.read",
         "crm.objects.deals.read",
         "crm.objects.deals.write",
+        "crm.objects.users.read",
+        "crm.objects.owners.read",
       ],
       ...rest
     } = props;