+++ packages/@runlightyear/linear/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @runlightyear/linear
 
+## 0.10.4
+
+### Patch Changes
+
+- Updated dependencies [e7dcf91c1]
+- Updated dependencies [e7dcf91c1]
+  - @runlightyear/lightyear@1.10.0
+
 ## 0.10.3
 
 ### Patch Changes