+++ packages/@runlightyear/postmark/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @runlightyear/postmark
 
+## 0.5.4
+
+### Patch Changes
+
+- Updated dependencies [e7dcf91c1]
+- Updated dependencies [e7dcf91c1]
+  - @runlightyear/lightyear@1.10.0
+
 ## 0.5.3
 
 ### Patch Changes