+++ packages/@runlightyear/slack/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @runlightyear/slack
 
+## 0.18.5
+
+### Patch Changes
+
+- Updated dependencies [e7dcf91c1]
+- Updated dependencies [e7dcf91c1]
+  - @runlightyear/lightyear@1.10.0
+
 ## 0.18.4
 
 ### Patch Changes