+++ packages/@runlightyear/smtp/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @runlightyear/smtp
 
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies [e7dcf91c1]
+- Updated dependencies [e7dcf91c1]
+  - @runlightyear/lightyear@1.10.0
+
 ## 0.3.3
 
 ### Patch Changes